Split navigation and consent surfaces
